Phase 14b: extract RNSPeerLifecycle
RNS.java 713 → 495 L. Everything that happens to a peer between "exists" and
"gone" — accepting an inbound link, creating an outbound one, and the ordered
teardown — moves to one class.

The split against RNSPeerRegistry is the point and is preserved exactly: the
registry mutates a list and republishes a snapshot under its own lock and hands
back whatever it displaced; RNSPeerLifecycle runs the side effects
(shutdownChannel, closeIfActive, makePeerUnavailable) with no registry lock
held. That is rule 4 of §0, and it now has a class boundary rather than a
comment holding it up.

Two changes beyond the move, both registered in §9:

- item 21: baseClientConnected and dataClientConnected collapse into one
  clientConnected(link, aspect). They were near-identical copies — the shape
  that hid §14.1 finding 1 — and the aspect was already the only difference.
  Their two INFO lines per connection become one carrying the same linkId and
  hash; the dropped half logged a raw byte[] toString next to its own hex.
- item 22: removePeer is deleted. Zero callers in src/main or src/test; it was
  package-private, so phase 1's sweep of zero-caller *public* methods missed it.

markPeerForImmediateRemoval and dedupIncomingPeerByIdentity keep three-line
delegators on RNS: ReticulumPeer reaches both through RNS.getInstance(), and §1
scopes ReticulumPeer out of this plan, so the call sites stay untouched.

The pruner and the lifecycle are now built in the constructor rather than as
field initialisers — both need rnsWorkerPool, which the constructor assigns.

All nine §10 comments grep-verified in their new homes. compile + test-compile
clean, 63 tests green.
